+++ creatures/creature.cpp
@@ -615,7 +615,7 @@ void Creature::onDeath() {
 
 	const int64_t timeNow = OTSYS_TIME();
 	const uint32_t inFightTicks = g_configManager().getNumber(PZ_LOCKED);
-	int32_t mostDamage = 0;
+	int64_t mostDamage = 0;
 	std::map<Creature*, uint64_t> experienceMap;
 	for (const auto &it : damageMap) {
 		if (Creature* attacker = g_game().getCreatureByID(it.first)) {
@@ -768,13 +768,13 @@ Item* Creature::getCorpse(Creature*, Creature*) {
 	return Item::CreateItem(getLookCorpse());
 }
 
-void Creature::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*/) {
-	int32_t oldHealth = health;
+void Creature::changeHealth(int64_t healthChange, bool sendHealthChange /* = true*/) {
+	int64_t oldHealth = health;
 
 	if (healthChange > 0) {
-		health += std::min<int32_t>(healthChange, getMaxHealth() - health);
+		health += std::min<int64_t>(healthChange, getMaxHealth() - health);
 	} else {
-		health = std::max<int32_t>(0, health + healthChange);
+		health = std::max<int64_t>(0, health + healthChange);
 	}
 
 	if (sendHealthChange && oldHealth != health) {
@@ -785,22 +785,22 @@ void Creature::changeHealth(int32_t healthChange, bool sendHealthChange /* = tru
 	}
 }
 
-void Creature::changeMana(int32_t manaChange) {
+void Creature::changeMana(int64_t manaChange) {
 	if (manaChange > 0) {
-		mana += std::min<int32_t>(manaChange, getMaxMana() - mana);
+		mana += std::min<int64_t>(manaChange, getMaxMana() - mana);
 	} else {
-		mana = std::max<int32_t>(0, mana + manaChange);
+		mana = std::max<int64_t>(0, mana + manaChange);
 	}
 }
 
-void Creature::gainHealth(Creature* healer, int32_t healthGain) {
+void Creature::gainHealth(Creature* healer, int64_t healthGain) {
 	changeHealth(healthGain);
 	if (healer) {
 		healer->onTargetCreatureGainHealth(this, healthGain);
 	}
 }
 
-void Creature::drainHealth(Creature* attacker, int32_t damage) {
+void Creature::drainHealth(Creature* attacker, int64_t damage) {
 	changeHealth(-damage, false);
 
 	if (attacker) {
@@ -808,7 +808,7 @@ void Creature::drainHealth(Creature* attacker, int32_t damage) {
 	}
 }
 
-void Creature::drainMana(Creature* attacker, int32_t manaLoss) {
+void Creature::drainMana(Creature* attacker, int64_t manaLoss) {
 	onAttacked();
 	changeMana(-manaLoss);
 
@@ -987,8 +987,8 @@ bool Creature::setFollowCreature(Creature* creature) {
 }
 
 double Creature::getDamageRatio(Creature* attacker) const {
-	uint32_t totalDamage = 0;
-	uint32_t attackerDamage = 0;
+	uint64_t totalDamage = 0;
+	uint64_t attackerDamage = 0;
 
 	for (const auto &it : damageMap) {
 		const CountBlock_t &cb = it.second;
@@ -1009,7 +1009,7 @@ uint64_t Creature::getGainedExperience(Creature* attacker) const {
 	return std::floor(getDamageRatio(attacker) * getLostExperience());
 }
 
-void Creature::addDamagePoints(Creature* attacker, int32_t damagePoints) {
+void Creature::addDamagePoints(Creature* attacker, int64_t damagePoints) {
 	if (damagePoints <= 0) {
 		return;
 	}
@@ -1090,7 +1090,7 @@ void Creature::onAttacked() {
 	//
 }
 
-void Creature::onAttackedCreatureDrainHealth(Creature* target, int32_t points) {
+void Creature::onAttackedCreatureDrainHealth(Creature* target, int64_t points) {
 	target->addDamagePoints(this, points);
 }
